C#/KataTests/SpinWordsTests.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace KataTests;

[TestFixture]
public class SpinWordsTests
{
    [Test]
    public static void Test1()
    {
        Assert.AreEqual("emocleW", SpinWords.Spin("Welcome"));
    }

    [Test]
    public static void Test2()
    {
        Assert.AreEqual("Hey wollef sroirraw", SpinWords.Spin("Hey fellow warriors"));
    }

    [Test]
    public static void Test3()
    {
        Assert.AreEqual("This is a test", SpinWords.Spin("This is a test"));
    }

    [Test]
    public static void Test4()
    {
        Assert.AreEqual("This is rehtona test", SpinWords.Spin("This is another test"));
    }

    [Test]
    public static void Test5()
    {
        Assert.AreEqual("You are tsomla to the last test", SpinWords.Spin("You are almost to the last test"));
    }

    [Test]
    public static void Test6()
    {
        Assert.AreEqual("Just gniddik ereht is llits one more", SpinWords.Spin("Just kidding there is still one more"));
    }
}

